refactor(rvr): update algebra, ecc, pairing and riscv to use 64bit address pointer

+++ extensions/riscv/rvr/c/rv64i_phantom_callbacks.h
@@ -8,7 +8,7 @@
  * Backed by a dispatch table installed by
  * `Rv64IExtension::register_host_callbacks`. */
 void openvm_hint_input(void);
-void openvm_print_str(uint32_t ptr, uint32_t len);
+void openvm_print_str(uint64_t ptr, uint32_t len);
 void openvm_hint_random(uint32_t num_words);
 
 /* Extension hint stream access (called by extension FFI staticlibs). */
